refactor: extract organization skill management
